# docs/plans/api-specification.yaml
openapi: 3.1.0
info:
  title: ekaiX AIXcelerator API
  version: 1.0.0
  description: |
    Backend API for ekaiX — a Snowflake Native App that enables business users to create
    semantic models and Cortex Agents through AI conversation.

    **Authentication:** All requests are authenticated via the `Sf-Context-Current-User` header
    injected by SPCS (Snowpark Container Services). No separate auth system exists.

    **Workspace Isolation:** All data is scoped to the authenticated user's workspace via
    PostgreSQL Row-Level Security. Sharing is explicit via the shares endpoints.

    **Streaming:** Agent responses are delivered via Server-Sent Events (SSE) on the
    `/agent/stream/{session_id}` endpoint.
  contact:
    name: ekai Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{spcs_endpoint}
    description: SPCS production
    variables:
      spcs_endpoint:
        default: ekaix.snowflakecomputing.app
...
security:
  - spcsAuth: []
components:
  securitySchemes:
    spcsAuth:
      type: apiKey
      in: header
      name: Sf-Context-Current-User
      description: Snowflake username injected by SPCS. Cannot be spoofed in production.

  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code (e.g., NOT_FOUND, VALIDATION_ERROR, FORBIDDEN)
          examples: ["NOT_FOUND"]
        message:
          type: string
          description: Human-readable error description
          examples: ["Data product not found"]
        details:
          type: object
          description: Additional error context (validation errors, field-level issues)
          additionalProperties: true
...
    AgentMessage:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: User message to the AI agent
          minLength: 1
          maxLength: 10000
        attachments:
          type: array
          items:
            type: object
            required: [type, id]
            properties:
              type:
                type: string
                enum: [document, artifact]
              id:
                type: string
                format: uuid
    AgentResponse:
      type: object
      required: [session_id, message_id, status]
      properties:
        session_id:
          type: string
          description: Agent session identifier (for SSE streaming)
        message_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [processing, completed, error]
    AgentStreamEvent:
      type: object
      description: |
        SSE event payload. Event types:
        - `token`: Streaming text token
        - `tool_call`: Agent is calling a tool
        - `tool_result`: Tool execution completed
        - `phase_change`: Agent moved to a new phase
        - `artifact`: New artifact generated (ERD, YAML, BRD)
        - `approval_request`: Agent needs user confirmation
        - `error`: Error occurred
        - `done`: Stream complete
      required: [type]
      properties:
        type:
          type: string
          enum: [token, tool_call, tool_result, phase_change, artifact, approval_request, error, done]
        data:
          oneOf:
            - type: object
              description: Token event
              properties:
                text:
                  type: string
            - type: object
              description: Tool call event
              properties:
                tool_name:
                  type: string
                tool_input:
                  type: object
            - type: object
              description: Phase change event
              properties:
                from_phase:
                  type: string
                to_phase:
                  type: string
            - type: object
              description: Artifact event
              properties:
                artifact_id:
                  type: string
                  format: uuid
                artifact_type:
                  type: string
                  enum: [erd, yaml, brd]
            - type: object
              description: Approval request event
              properties:
                action:
                  type: string
                description:
                  type: string
                options:
                  type: array
                  items:
                    type: string
            - type: object
              description: Error event
              properties:
                error:
                  type: string
                message:
                  type: string
...
paths:
...
  /agent/message:
    post:
      tags: [Agent]
      operationId: sendAgentMessage
      summary: Send a message to the AI agent
      description: |
        Sends a user message to the Deep Agents orchestrator. Returns a session_id
        for subscribing to the SSE stream. The agent processes the message asynchronously
        and streams responses via `/agent/stream/{session_id}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data_product_id, message]
              properties:
                data_product_id:
                  type: string
                  format: uuid
                message:
                  $ref: "#/components/schemas/AgentMessage"
      responses:
        "200":
          description: Message accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponse"
        "400":
          description: Invalid message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Data product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agent/stream/{session_id}:
    get:
      tags: [Agent]
      operationId: streamAgentResponse
      summary: Subscribe to agent response stream (SSE)
      description: |
        Server-Sent Events stream for real-time agent responses. Connect after
        sending a message via `/agent/message`.

        **Event format:**
        ```
        event: token
        data: {"text": "Let me analyze..."}

        event: tool_call
        data: {"tool_name": "query_information_schema", "tool_input": {...}}

        event: artifact
        data: {"artifact_id": "uuid", "artifact_type": "erd"}

        event: done
        data: {}
        ```

        The stream ends with a `done` event. Client should reconnect only after
        sending a new message.
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/AgentStreamEvent"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...